backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "av==17.0.0",
    "certifi==2026.2.25",
    "click==8.3.1",
    "colorama==0.4.6",
    "ctranslate2==4.7.1",
    "diskcache==5.6.3",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi==0.135.2",
    "fastapi-cli==0.0.24",
    "fastapi-cloud-cli==0.15.1",
    "fastar==0.9.0",
    "faster-whisper==1.2.1",
    "filelock==3.25.2",
    "flatbuffers==25.12.19",
    "fsspec==2026.3.0",
    "h11==0.16.0",
    "hf-xet==1.4.2",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "huggingface-hub==1.8.0",
    "idna==3.11",
    "jinja2==3.1.6",
    "llama-cpp-agent>=0.2.35",
    "llama-cpp-python>=0.3.19",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.4",
    "mpmath==1.3.0",
    "numpy==2.4.3",
    "onnx==1.21.0",
    "onnxruntime==1.24.4",
    "openai>=2.30.0",
    "outlines[llamacpp]>=1.2.12",
    "packaging==26.0",
    "pathvalidate==3.3.1",
    "piper-tts==1.4.1",
    "protobuf==7.34.1",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-extra-types==2.11.1",
    "pydantic-settings==2.13.1",
    "pygments==2.19.2",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.22",
    "pyyaml==6.0.3",
    "rich==14.3.3",
    "rich-toolkit==0.19.7",
    "rignore==0.7.6",
    "sentry-sdk==2.56.0",
    "setuptools==82.0.1",
    "shellingham==1.5.4",
    "starlette==1.0.0",
    "sympy==1.14.0",
    "tokenizers==0.22.2",
    "tqdm==4.67.3",
    "typer==0.24.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uvicorn==0.42.0",
    "watchfiles==1.1.1",
    "websockets==16.0",
]
